+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change history for ui-users
 
-## 5.0.0 (IN PROGRESS)
+## [5.0.0](https://github.com/folio-org/ui-users/tree/v5.0.0) (2020-10-14)
+[Full Changelog](https://github.com/folio-org/ui-users/compare/v4.0.0...v5.0.0)
 
 * modified the lost date display to include aged to lost status. Fixes UIU-1804.
 * Fixing breaking tests.